+++ ui/src/utils/queryUrlGenerator.js
@@ -1,11 +1,12 @@
 import AJAX from 'utils/ajax'
 
-export const proxy = async ({source, query, db, rp}) => {
+export const proxy = async ({source, query, db, rp, templates}) => {
   try {
     return await AJAX({
       method: 'POST',
       url: source,
       data: {
+        templates,
         query,
         db,
         rp,